+++ Starstorm 2/Cores/NemesisInvasion/Components/NemesisInvasionManager.cs
@@ -23,6 +23,9 @@ namespace Starstorm2Unofficial.Cores.NemesisInvasion.Components
         public static HashSet<ItemIndex> itemWhitelist = new HashSet<ItemIndex>();
         public static bool allowSpawnInBazaar = false;
         private static SceneDef bazaarSceneDef = Addressables.LoadAssetAsync<SceneDef>("RoR2/Base/bazaar/bazaar.asset").WaitForCompletion();
+        private static SceneDef arenaSceneDef = Addressables.LoadAssetAsync<SceneDef>("RoR2/Base/arena/arena.asset").WaitForCompletion();
+
+        public static bool requireFullVoid = true;
 
         private bool playedEventStartChatMessage = false;
         private bool playedEventEndChatMessage = false;
@@ -77,11 +80,19 @@ namespace Starstorm2Unofficial.Cores.NemesisInvasion.Components
             if (NemesisInvasionManager.instance)
             {
                 SceneDef sd = SceneCatalog.GetSceneDefForCurrentScene();
-                if (sd && !sd.isFinalStage)
+                if (sd)
                 {
-                    if (NemesisInvasionManager.instance.voidClearedSuccessfully)
+                    if (!sd.isFinalStage && sd != arenaSceneDef)
                     {
-                        NemesisInvasionManager.instance.SpawnNemesis();
+                        if (NemesisInvasionManager.instance.voidClearedSuccessfully)
+                        {
+                            NemesisInvasionManager.instance.SpawnNemesis();
+                        }
+                    }
+
+                    if (!requireFullVoid && sd == arenaSceneDef)
+                    {
+                        NemesisInvasionManager.instance.voidClearedSuccessfully = true;
                     }
                 }
             }